clean up common code into concern
