+++ app/models/genre.rb
@@ -1,8 +1,8 @@
 class Genre < ActiveRecord::Base
+  include HasVideo
+
   has_many :genres_videos
-  has_many :videos, through: :genres_videos
-  has_many :movies, -> { where(videos: {type: 'Movie'}) }, through: :genres_videos, source: :video
-  has_many :tv_shows, -> { where(videos: {type: 'TvShow'}) }, through: :genres_videos, source: :video
+  has_videos(through: :genres_videos)
 
   before_validation :sanitize_name
   validates :name, presence: true, uniqueness: {case_sensitive: false}
@@ -11,18 +11,6 @@ class Genre < ActiveRecord::Base
     "sci fi" => "Science Fiction",
     "scifi" => "Science Fiction"
   }.freeze
-
-  def has_movies?
-    movies.count > 0
-  end
-
-  def has_local_movies?
-    movies.local.count > 0
-  end
-
-  def has_tv_shows?
-    tv_shows.count > 0
-  end
 
   def self.normalized_name(name)
     name = name.strip.titleize
